requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
Deprecated==1.2.14
Django==3.2.21
frozenlist==1.3.3
future==0.18.3
gunicorn==21.2.0
idna==3.4
line-bot-sdk==3.1.0
lxml==4.9.3
multidict==6.0.4
numpy==1.21.6
packaging==23.1
pandas==1.3.5
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
yarl==1.9.2
